Add list for digital objects to test preview image code with existing images
