+++ dc-cudami-admin/src/main/resources/messages.properties
@@ -106,10 +106,12 @@ lastname=Lastname
 lbl.actions=Actions
 lbl.article.inline=article
 lbl.created=created
+lbl.digital-object.inline=the digital object
 lbl.itemlocale=Show language dependent values in
 lbl.label=Label
 lbl.last_modified=last modified
-lbl.predicate.inline=relation type
+lbl.predicate.inline=the relation type
+lbl.preview_image=Preview Image
 lbl.row_number=#
 lbl.uuid=UUID
 lbl.value=Value
@@ -145,6 +147,7 @@ number_of_binary_resources=Number of files
 objects=Objects
 ok=Ok
 page.title.articles=Articles
+page.title.digital_objects=Digital Objects
 parent_collections=Parent collections
 parent_digital_object=Enclosing digital object
 parent_item=Parent item
@@ -191,6 +194,7 @@ title_warning=Warning
 to=to
 tooltip.delete=delete
 tooltip.edit=edit
+tooltip.no_preview_image=no preview image
 tooltip.view=view
 tooltip_activate_user=activate user
 tooltip_currently_published=currently published
