+++ dc-cudami-admin/src/main/resources/messages_de.properties
@@ -106,10 +106,12 @@ lastname=Nachname
 lbl.actions=Aktionen
 lbl.article.inline=Artikel
 lbl.created=erstellt
+lbl.digital-object.inline=das Digitalisat (Digitale Objekt)
 lbl.itemlocale=Sprachspezifische Werte anzeigen in
 lbl.label=Bezeichnung
 lbl.last_modified=letzte Änderung
-lbl.predicate.inline=Beziehungstyp
+lbl.predicate.inline=den Beziehungstyp
+lbl.preview_image=Vorschaubild
 lbl.row_number=#
 lbl.uuid=UUID
 lbl.value=Wert
@@ -145,6 +147,7 @@ number_of_binary_resources=Anzahl der Dateien
 objects=Objekte
 ok=Ok
 page.title.articles=Artikel
+page.title.digital_objects=Digitalisate (digitale Objekte)
 parent_collections=Übergeordnete Sammlungen
 parent_digital_object=Umschließendes Digitalisat (Digitales Objekt)
 parent_item=Übergeordnetes Exemplar
@@ -191,6 +194,7 @@ title_warning=Warnung
 to=bis
 tooltip.delete=löschen
 tooltip.edit=bearbeiten
+tooltip.no_preview_image=kein Vorschaubild
 tooltip.view=anzeigen
 tooltip_activate_user=Benutzer aktivieren
 tooltip_currently_published=aktuell veröffentlicht
